Add CSV export button to works
Closes #90
