+++ app/client/components/work.less
@@ -38,6 +38,10 @@ av-work {
   &.Work-headerButton--autoColorize {
     .m-icon-paint-brush;
   }
+
+  &.Work-headerButton--exportCsv {
+    .m-icon-file-excel-o;
+  }
 }
 
 .Work-body {
